requirements.txt
python==3.10.12
Flask
pytorch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
pytorch-transformers==1.2.0
huggingface_hub[inference]
transformers -U
importlib_metadata
easydict